Added some code to serve as the skeleton
The code consists of some constants that may be important later on, a
few skeletal class definitions, and some basic framework for the
graphical display with very minimal interaction.  Just trying to get
some of the initial work out of the way so it will be easier to build
from here.