Shuffling and beginning of player selection
Added shuffling functionality to be used to reorder the card decks
whenever necessary.  Started to work on the screen that will allow
players to select how many human players (anywhere from 1 to 4, I
believe from what we discussed in class that the game will always have 4
players total with n being human and (4 - n) being AI).

It will take a bit of work to build the whole GUI for having buttons
selected and everything but the good thing is that once that is built we
should very easily be able to build other things, like a title screen or
options screen, based off of this code.

Also fixed a few spelling errors.