Added a .gitignore file and removed .idea directory
Due to my using of the PyCharm IDE with the workspace set to the local
directory of the repository on my computer, there was a directory titled
".idea" that contained a lot of files useful to myself in development
but that did not necessarily belong in the repository itself and would
have served no use to any other members of the group.

A .gitignore file was also added in order to prevent any further changes
in this directory to be pushed to the repository, and this can be
updated at a later date with any extra directories that need to be
excluded, such as potentially the workspace directories of other group
members' IDEs.