TextRect
Added a python module called TextRect, which is used for rendering
wordwrapped text.  We will likely use this to display text on the action
and scenario cards.  It is made to work with pygame.font

Also added a few new graphics and an update to the README including a
citation regarding TextRect.