Rearrangements and real dealing
I rearranged a bunch of the code to make it work better and made the
whole cards being dealt to a player's hand thing actually work.  Before
it had been actually just accessing them directly from the deck.

I also added the scenario card to the center of the screen and it gets
dealt a new scenario each time.

The current major issue is I'm not really sure how we're gonna handle
multiple player stuff.  I think I'm just not arranging nested loops in a
way that's anywhere near correct because every attempt to do so seems to
end with an infinite loop during which I can't even close the game.