Working on Customization Screen
Significant progress made towards the customization screen, including
rendering of graphics, creation of sprites and objects, and mapping of
coordinates.  Some other stuff was changed including formatting of
strings and rearranging of a few lines for continuity purposes.

Next work to be done is to actually write the code to display the
customization screen and then to give it functionality.