Customization screen works!
Built the customization screen and gave it functionality.  Card back
that is selected is stored in a string called "cardArtwork".

At the time being the way that selecting this stuff, specifically how
image changes are dealt with, isn't super efficient.  If we have
problems with that down the line, we can use that same array method that
was used for selecting cards in the actual in-game stuff.