Added more functionality to unfinished screens
Instructions and About screens now have main menu buttons.  About screen
can now be correctly accessed.  Basic framework for diagnostics screen
started, and the diagnostics screen can be accessed by pressing the 'D'
key on the main menu.

When on the diagnostics screen, a screenshot can be created and saved to
the game directory as "diagnosticOutput.png" by pressing the 'P' key.