Incremental Difficulty
Basic incremental difficulty system created.

"Winning streak" is tracked as number of rounds in a row that player has
currently gotten any nonzero number of points on.

Amount of time given for the round in seconds is equal to the base time
for that difficulty divided by (1 + (winningStreak / 10)).

So, a winning streak of 10 in a row effectively gives half the time to
answer.

This could all easily be changed down the line but it is good to have it
totally functional.